test(verify): isolate parameter safety diagnostics
